Phase 3: extract ToolCardShell and host View raw on the shell
Extract the shared approval chrome (header, collapse, decision buttons, result
section, result-review callout) into tool_renderers/tool_card_shell.tsx, and
move the View raw affordance onto the shell so rich cards inherit it. ToolCard
becomes a thin shell + ToolArguments field-list body. ToolArguments keeps the
field list + Show more; its raw view is exposed via ToolArgumentsRaw.
